requirements.txt
streamlitstreamlit
pandas
plotly
feedparser